requirements.txt
#h5py==3.12.1
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
jsonschema==4.23.0
#setuptools==75.2.0
#wheel

yt_dlp==2025.3.31

Hypercorn==0.17.3
flask==3.0.3
flask_session==0.8.0
redis==5.0.8
paho-mqtt==2.1.0
authlib==1.3.2
grafana-client==4.3.2
Flask-Cors==4.0.1
cryptography==43.0.1

shapely>=2.0.0

psutil==6.0.0
gputil==1.4.0

#https://github.com/ultralytics/ultralytics/issues/14235
tensorflow==2.16.1
tf-keras==2.16.0
onnx2tf==1.22.3

ultralytics==8.3.16
